requirements.txt
flask
cassandra-driver
redis
streamlit
requests
pytest